requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Bcrypt
python-dotenv
psycopg2-binary
Flask-JWT-Extended
requests
Flask-Cors
marshmallow
gunicorn